[Update] Constructor of binarygen to make it slightly easier to do unit test

+++ inc/binaryplink.hpp
@@ -41,7 +41,7 @@ public:
      * \param reporter is the logger
      */
     BinaryPlink(const std::string& file_list, const std::string& file,
-                const uint32_t thread, const bool ignore_fid,
+                const size_t thread, const bool ignore_fid,
                 const bool keep_nonfounder, const bool keep_ambig,
                 const bool is_ref, Reporter& reporter);
     BinaryPlink() {}
